[service] Move PackageCallbacks to Software::Callbacks::Progress
